Move `FeedLoaderPresentationAdapter` to it's own file
